backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
twilio==8.12.0
openpyxl==3.1.2
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
drf-spectacular==0.26.3
python-json-logger==2.0.7
Pillow==10.1.0
requests==2.31.0
reportlab==4.0.7
urllib3==2.0.7
cryptography==42.0.5
requests-oauthlib==1.3.1
pyotp==2.9.0
qrcode==7.4.2
django-templated-mail==1.1.1
zxcvbn==4.4.28
django-ratelimit==3.0.1
django-csp==4.0
fcm-django==2.0.0
firebase-admin==6.1.0
python-decouple==3.6
celery==5.2.7
redis==4.5.3
croniter==1.4.1
xlsxwriter==3.1.9
django-extensions==3.2.3
drf-spectacular-sidecar==2024.7.1
django-filter==23.5
tenacity==8.2.3
setuptools<81

# Production Payment Providers
stripe==7.9.0


# Additional Production Dependencies
gunicorn==21.2.0
uvicorn[standard]==0.27.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Celery Beat with DB scheduler
django-celery-beat==2.5.0

# Monitoring and Logging
prometheus-client==0.19.0
django-prometheus==2.3.1

# Security
django-axes==6.3.0

# Social Authentication
django-allauth==0.63.3

# Redis cache
django-redis==5.4.0

# Email Service
sendgrid==6.11.0
